Moving UserDefaults persistence to be in VFoundation
